data streaming fix
This is a temporary fix that fixes the issue where we were unable to
post large strings (node refused them). Rather than change node’s
default parameters, we switched to posting smaller packets — at the
episode level. In the future this should be changed to a streaming
socket, but that is beyond my ability at the present time.

Additionally, we fixed an issue whereby, if you refreshed the page it
would not start a new file, it would just tack on to the end of the
file that was already started.

Lastly, we added a server timeout. However, we need to test that this
is a sufficient duration for our purposes.